HBASE-24986 Move ReplicationBarrier related methods to a separated class (#2354)
