+++ hbase-server/src/main/java/org/apache/hadoop/hbase/master/CatalogJanitor.java
@@ -189,7 +189,7 @@ public class CatalogJanitor extends ScheduledChore {
           break;
         }
 
-        List<RegionInfo> parents = MetaTableAccessor.getMergeRegions(e.getValue().rawCells());
+        List<RegionInfo> parents = CatalogFamilyFormat.getMergeRegions(e.getValue().rawCells());
         if (parents != null && cleanMergeRegion(e.getKey(), parents)) {
           gcs++;
         }
@@ -323,7 +323,7 @@ public class CatalogJanitor extends ScheduledChore {
   boolean cleanParent(final RegionInfo parent, Result rowContent)
   throws IOException {
     // Check whether it is a merged region and if it is clean of references.
-    if (MetaTableAccessor.hasMergeRegions(rowContent.rawCells())) {
+    if (CatalogFamilyFormat.hasMergeRegions(rowContent.rawCells())) {
       // Wait until clean of merge parent regions first
       return false;
     }
@@ -580,7 +580,7 @@ public class CatalogJanitor extends ScheduledChore {
         if (regionInfo.isSplitParent()) { // splitParent means split and offline.
           this.report.splitParents.put(regionInfo, r);
         }
-        if (MetaTableAccessor.hasMergeRegions(r.rawCells())) {
+        if (CatalogFamilyFormat.hasMergeRegions(r.rawCells())) {
           this.report.mergedRegions.put(regionInfo, r);
         }
       }
