+++ hbase-server/src/main/java/org/apache/hadoop/hbase/master/assignment/GCMultipleMergedRegionsProcedure.java
@@ -20,8 +20,6 @@ package org.apache.hadoop.hbase.master.assignment;
 import java.io.IOException;
 import java.util.List;
 import java.util.stream.Collectors;
-
-import org.apache.hadoop.hbase.MetaTableAccessor;
 import org.apache.hadoop.hbase.TableName;
 import org.apache.hadoop.hbase.client.RegionInfo;
 import org.apache.hadoop.hbase.master.procedure.AbstractStateMachineTableProcedure;
@@ -29,12 +27,12 @@ import org.apache.hadoop.hbase.master.procedure.MasterProcedureEnv;
 import org.apache.hadoop.hbase.procedure2.ProcedureStateSerializer;
 import org.apache.hadoop.hbase.procedure2.ProcedureSuspendedException;
 import org.apache.hadoop.hbase.procedure2.ProcedureYieldException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import org.apache.hadoop.hbase.shaded.protobuf.ProtobufUtil;
 import org.apache.hadoop.hbase.shaded.protobuf.generated.MasterProcedureProtos.GCMergedRegionsState;
 import org.apache.hadoop.hbase.shaded.protobuf.generated.MasterProcedureProtos.GCMultipleMergedRegionsStateData;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 /**
  * GC regions that have been Merged. Caller determines if it is GC time. This Procedure does not
@@ -100,8 +98,8 @@ public class GCMultipleMergedRegionsProcedure extends
           // If GCMultipleMergedRegionsProcedure processing is slower than the CatalogJanitor's scan
           // interval, it will end resubmitting GCMultipleMergedRegionsProcedure for the same
           // region. We can skip duplicate GCMultipleMergedRegionsProcedure while previous finished
-          List<RegionInfo> parents = MetaTableAccessor.getMergeRegions(
-            env.getMasterServices().getConnection(), mergedChild.getRegionName());
+          List<RegionInfo> parents =
+            env.getAssignmentManager().getRegionStateStore().getMergeRegions(mergedChild);
           if (parents == null || parents.isEmpty()) {
             LOG.info("{} mergeXXX qualifiers have ALL been deleted", mergedChild.getShortNameToLog());
             return Flow.NO_MORE_STATE;
@@ -113,8 +111,7 @@ public class GCMultipleMergedRegionsProcedure extends
           setNextState(GCMergedRegionsState.GC_REGION_EDIT_METADATA);
           break;
         case GC_REGION_EDIT_METADATA:
-          MetaTableAccessor.deleteMergeQualifiers(env.getMasterServices().getConnection(),
-              mergedChild);
+          env.getAssignmentManager().getRegionStateStore().deleteMergeQualifiers(mergedChild);
           return Flow.NO_MORE_STATE;
         default:
           throw new UnsupportedOperationException(this + " unhandled state=" + state);
