+++ hbase-server/src/main/java/org/apache/hadoop/hbase/master/assignment/RegionStateStore.java
@@ -18,33 +18,48 @@
 package org.apache.hadoop.hbase.master.assignment;
 
 import java.io.IOException;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
-import java.util.SortedMap;
-import java.util.TreeMap;
+import java.util.concurrent.CompletableFuture;
+import java.util.stream.Collectors;
 import org.apache.hadoop.hbase.CatalogFamilyFormat;
 import org.apache.hadoop.hbase.Cell;
+import org.apache.hadoop.hbase.Cell.Type;
 import org.apache.hadoop.hbase.CellBuilderFactory;
 import org.apache.hadoop.hbase.CellBuilderType;
+import org.apache.hadoop.hbase.CellUtil;
 import org.apache.hadoop.hbase.ClientMetaTableAccessor;
+import org.apache.hadoop.hbase.DoNotRetryIOException;
 import org.apache.hadoop.hbase.HConstants;
 import org.apache.hadoop.hbase.HRegionLocation;
 import org.apache.hadoop.hbase.MetaTableAccessor;
 import org.apache.hadoop.hbase.RegionLocations;
 import org.apache.hadoop.hbase.ServerName;
 import org.apache.hadoop.hbase.TableName;
+import org.apache.hadoop.hbase.client.AsyncTable;
+import org.apache.hadoop.hbase.client.Delete;
+import org.apache.hadoop.hbase.client.Get;
+import org.apache.hadoop.hbase.client.Mutation;
 import org.apache.hadoop.hbase.client.Put;
 import org.apache.hadoop.hbase.client.RegionInfo;
+import org.apache.hadoop.hbase.client.RegionInfoBuilder;
+import org.apache.hadoop.hbase.client.RegionReplicaUtil;
 import org.apache.hadoop.hbase.client.Result;
 import org.apache.hadoop.hbase.client.Table;
 import org.apache.hadoop.hbase.client.TableDescriptor;
 import org.apache.hadoop.hbase.master.MasterFileSystem;
 import org.apache.hadoop.hbase.master.MasterServices;
+import org.apache.hadoop.hbase.master.RegionState;
 import org.apache.hadoop.hbase.master.RegionState.State;
 import org.apache.hadoop.hbase.procedure2.Procedure;
 import org.apache.hadoop.hbase.procedure2.util.StringUtils;
+import org.apache.hadoop.hbase.replication.ReplicationBarrierFamilyFormat;
 import org.apache.hadoop.hbase.util.Bytes;
 import org.apache.hadoop.hbase.util.EnvironmentEdgeManager;
+import org.apache.hadoop.hbase.util.FutureUtils;
 import org.apache.hadoop.hbase.wal.WALSplitUtil;
 import org.apache.hadoop.hbase.zookeeper.MetaTableLocator;
 import org.apache.yetus.audience.InterfaceAudience;
@@ -52,7 +67,14 @@ import org.apache.zookeeper.KeeperException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import org.apache.hbase.thirdparty.com.google.common.annotations.VisibleForTesting;
 import org.apache.hbase.thirdparty.com.google.common.base.Preconditions;
+
+import org.apache.hadoop.hbase.shaded.protobuf.ProtobufUtil;
+import org.apache.hadoop.hbase.shaded.protobuf.generated.ClientProtos;
+import org.apache.hadoop.hbase.shaded.protobuf.generated.MultiRowMutationProtos.MultiRowMutationService;
+import org.apache.hadoop.hbase.shaded.protobuf.generated.MultiRowMutationProtos.MutateRowsRequest;
+import org.apache.hadoop.hbase.shaded.protobuf.generated.MultiRowMutationProtos.MutateRowsResponse;
 
 /**
  * Store Region State to hbase:meta table.
@@ -60,6 +82,7 @@ import org.apache.hbase.thirdparty.com.google.common.base.Preconditions;
 @InterfaceAudience.Private
 public class RegionStateStore {
   private static final Logger LOG = LoggerFactory.getLogger(RegionStateStore.class);
+  private static final Logger METALOG = LoggerFactory.getLogger("org.apache.hadoop.hbase.META");
 
   /** The delimiter for meta columns for replicaIds &gt; 0 */
   protected static final char META_REPLICA_ID_DELIMITER = '_';
@@ -190,7 +213,7 @@ public class RegionStateStore {
       // only update replication barrier for default replica
       if (regionInfo.getReplicaId() == RegionInfo.DEFAULT_REPLICA_ID &&
         hasGlobalReplicationScope(regionInfo.getTable())) {
-        MetaTableAccessor.addReplicationBarrier(put, openSeqNum);
+        ReplicationBarrierFamilyFormat.addReplicationBarrier(put, openSeqNum);
         info.append(", repBarrier=").append(openSeqNum);
       }
       info.append(", openSeqNum=").append(openSeqNum);
@@ -242,33 +265,226 @@ public class RegionStateStore {
     return maxSeqId > 0 ? maxSeqId + 1 : HConstants.NO_SEQNUM;
   }
 
+  /**
+   * Performs an atomic multi-mutate operation against the given table. Used by the likes of merge
+   * and split as these want to make atomic mutations across multiple rows.
+   */
+  private void multiMutate(RegionInfo ri, List<Mutation> mutations) throws IOException {
+    debugLogMutations(mutations);
+    byte[] row =
+      Bytes.toBytes(RegionReplicaUtil.getRegionInfoForDefaultReplica(ri).getRegionNameAsString() +
+        HConstants.DELIMITER);
+    MutateRowsRequest.Builder builder = MutateRowsRequest.newBuilder();
+    for (Mutation mutation : mutations) {
+      if (mutation instanceof Put) {
+        builder.addMutationRequest(
+          ProtobufUtil.toMutation(ClientProtos.MutationProto.MutationType.PUT, mutation));
+      } else if (mutation instanceof Delete) {
+        builder.addMutationRequest(
+          ProtobufUtil.toMutation(ClientProtos.MutationProto.MutationType.DELETE, mutation));
+      } else {
+        throw new DoNotRetryIOException(
+          "multi in MetaEditor doesn't support " + mutation.getClass().getName());
+      }
+    }
+    MutateRowsRequest request = builder.build();
+    AsyncTable<?> table =
+      master.getConnection().toAsyncConnection().getTable(TableName.META_TABLE_NAME);
+    CompletableFuture<MutateRowsResponse> future =
+      table.<MultiRowMutationService, MutateRowsResponse> coprocessorService(
+        MultiRowMutationService::newStub,
+        (stub, controller, done) -> stub.mutateRows(controller, request, done), row);
+    FutureUtils.get(future);
+  }
+
+  private Result getRegionCatalogResult(RegionInfo region) throws IOException {
+    Get get =
+      new Get(CatalogFamilyFormat.getMetaKeyForRegion(region)).addFamily(HConstants.CATALOG_FAMILY);
+    try (Table table = master.getConnection().getTable(TableName.META_TABLE_NAME)) {
+      return table.get(get);
+    }
+  }
+
+
+  private static Put addSequenceNum(Put p, long openSeqNum, int replicaId) throws IOException {
+    return p.add(CellBuilderFactory.create(CellBuilderType.SHALLOW_COPY).setRow(p.getRow())
+      .setFamily(HConstants.CATALOG_FAMILY)
+      .setQualifier(CatalogFamilyFormat.getSeqNumColumn(replicaId)).setTimestamp(p.getTimestamp())
+      .setType(Type.Put).setValue(Bytes.toBytes(openSeqNum)).build());
+  }
+
   // ============================================================================================
   //  Update Region Splitting State helpers
   // ============================================================================================
-  public void splitRegion(RegionInfo parent, RegionInfo hriA, RegionInfo hriB,
-      ServerName serverName) throws IOException {
-    TableDescriptor htd = getDescriptor(parent.getTable());
+  /**
+   * Splits the region into two in an atomic operation. Offlines the parent region with the
+   * information that it is split into two, and also adds the daughter regions. Does not add the
+   * location information to the daughter regions since they are not open yet.
+   */
+  public void splitRegion(RegionInfo parent, RegionInfo splitA, RegionInfo splitB,
+    ServerName serverName, TableDescriptor htd) throws IOException {
     long parentOpenSeqNum = HConstants.NO_SEQNUM;
     if (htd.hasGlobalReplicationScope()) {
       parentOpenSeqNum = getOpenSeqNumForParentRegion(parent);
     }
-    MetaTableAccessor.splitRegion(master.getConnection(), parent, parentOpenSeqNum, hriA, hriB,
-      serverName, getRegionReplication(htd));
+    long time = EnvironmentEdgeManager.currentTime();
+    // Put for parent
+    Put putParent = MetaTableAccessor.makePutFromRegionInfo(
+      RegionInfoBuilder.newBuilder(parent).setOffline(true).setSplit(true).build(), time);
+    MetaTableAccessor.addDaughtersToPut(putParent, splitA, splitB);
+
+    // Puts for daughters
+    Put putA = MetaTableAccessor.makePutFromRegionInfo(splitA, time);
+    Put putB = MetaTableAccessor.makePutFromRegionInfo(splitB, time);
+    if (parentOpenSeqNum > 0) {
+      ReplicationBarrierFamilyFormat.addReplicationBarrier(putParent, parentOpenSeqNum);
+      ReplicationBarrierFamilyFormat.addReplicationParent(putA, Collections.singletonList(parent));
+      ReplicationBarrierFamilyFormat.addReplicationParent(putB, Collections.singletonList(parent));
+    }
+    // Set initial state to CLOSED
+    // NOTE: If initial state is not set to CLOSED then daughter regions get added with the
+    // default OFFLINE state. If Master gets restarted after this step, start up sequence of
+    // master tries to assign these offline regions. This is followed by re-assignments of the
+    // daughter regions from resumed {@link SplitTableRegionProcedure}
+    MetaTableAccessor.addRegionStateToPut(putA, RegionState.State.CLOSED);
+    MetaTableAccessor.addRegionStateToPut(putB, RegionState.State.CLOSED);
+
+    // new regions, openSeqNum = 1 is fine.
+    addSequenceNum(putA, 1, splitA.getReplicaId());
+    addSequenceNum(putB, 1, splitB.getReplicaId());
+
+    // Add empty locations for region replicas of daughters so that number of replicas can be
+    // cached whenever the primary region is looked up from meta
+    int regionReplication = getRegionReplication(htd);
+    for (int i = 1; i < regionReplication; i++) {
+      MetaTableAccessor.addEmptyLocation(putA, i);
+      MetaTableAccessor.addEmptyLocation(putB, i);
+    }
+
+    multiMutate(parent, Arrays.asList(putParent, putA, putB));
   }
 
   // ============================================================================================
   //  Update Region Merging State helpers
   // ============================================================================================
-  public void mergeRegions(RegionInfo child, RegionInfo [] parents, ServerName serverName)
-      throws IOException {
-    TableDescriptor htd = getDescriptor(child.getTable());
+  public void mergeRegions(RegionInfo child, RegionInfo[] parents, ServerName serverName,
+    TableDescriptor htd) throws IOException {
     boolean globalScope = htd.hasGlobalReplicationScope();
-    SortedMap<RegionInfo, Long> parentSeqNums = new TreeMap<>();
+    long time = HConstants.LATEST_TIMESTAMP;
+    List<Mutation> mutations = new ArrayList<>();
+    List<RegionInfo> replicationParents = new ArrayList<>();
     for (RegionInfo ri: parents) {
-      parentSeqNums.put(ri, globalScope? getOpenSeqNumForParentRegion(ri): -1);
+      long seqNum = globalScope ? getOpenSeqNumForParentRegion(ri) : -1;
+      // Deletes for merging regions
+      mutations.add(MetaTableAccessor.makeDeleteFromRegionInfo(ri, time));
+      if (seqNum > 0) {
+        mutations
+          .add(ReplicationBarrierFamilyFormat.makePutForReplicationBarrier(ri, seqNum, time));
+        replicationParents.add(ri);
+      }
     }
-    MetaTableAccessor.mergeRegions(master.getConnection(), child, parentSeqNums,
-        serverName, getRegionReplication(htd));
+    // Put for parent
+    Put putOfMerged = MetaTableAccessor.makePutFromRegionInfo(child, time);
+    putOfMerged = addMergeRegions(putOfMerged, Arrays.asList(parents));
+    // Set initial state to CLOSED.
+    // NOTE: If initial state is not set to CLOSED then merged region gets added with the
+    // default OFFLINE state. If Master gets restarted after this step, start up sequence of
+    // master tries to assign this offline region. This is followed by re-assignments of the
+    // merged region from resumed {@link MergeTableRegionsProcedure}
+    MetaTableAccessor.addRegionStateToPut(putOfMerged, RegionState.State.CLOSED);
+    mutations.add(putOfMerged);
+    // The merged is a new region, openSeqNum = 1 is fine. ServerName may be null
+    // if crash after merge happened but before we got to here.. means in-memory
+    // locations of offlined merged, now-closed, regions is lost. Should be ok. We
+    // assign the merged region later.
+    if (serverName != null) {
+      MetaTableAccessor.addLocation(putOfMerged, serverName, 1, child.getReplicaId());
+    }
+
+    // Add empty locations for region replicas of the merged region so that number of replicas
+    // can be cached whenever the primary region is looked up from meta
+    int regionReplication = getRegionReplication(htd);
+    for (int i = 1; i < regionReplication; i++) {
+      MetaTableAccessor.addEmptyLocation(putOfMerged, i);
+    }
+    // add parent reference for serial replication
+    if (!replicationParents.isEmpty()) {
+      ReplicationBarrierFamilyFormat.addReplicationParent(putOfMerged, replicationParents);
+    }
+    multiMutate(child, mutations);
+  }
+
+  /**
+   * Check whether the given {@code region} has any 'info:merge*' columns.
+   */
+  public boolean hasMergeRegions(RegionInfo region) throws IOException {
+    return CatalogFamilyFormat.hasMergeRegions(getRegionCatalogResult(region).rawCells());
+  }
+
+  /**
+   * @return Return all regioninfos listed in the 'info:merge*' columns of the given {@code region}.
+   */
+  public List<RegionInfo> getMergeRegions(RegionInfo region) throws IOException {
+    return CatalogFamilyFormat.getMergeRegions(getRegionCatalogResult(region).rawCells());
+  }
+
+  /**
+   * Deletes merge qualifiers for the specified merge region.
+   * @param connection connection we're using
+   * @param mergeRegion the merged region
+   */
+  public void deleteMergeQualifiers(RegionInfo mergeRegion)
+    throws IOException {
+    // NOTE: We are doing a new hbase:meta read here.
+    Cell[] cells = getRegionCatalogResult(mergeRegion).rawCells();
+    if (cells == null || cells.length == 0) {
+      return;
+    }
+    Delete delete = new Delete(mergeRegion.getRegionName());
+    List<byte[]> qualifiers = new ArrayList<>();
+    for (Cell cell : cells) {
+      if (!CatalogFamilyFormat.isMergeQualifierPrefix(cell)) {
+        continue;
+      }
+      byte[] qualifier = CellUtil.cloneQualifier(cell);
+      qualifiers.add(qualifier);
+      delete.addColumns(HConstants.CATALOG_FAMILY, qualifier, HConstants.LATEST_TIMESTAMP);
+    }
+
+    // There will be race condition that a GCMultipleMergedRegionsProcedure is scheduled while
+    // the previous GCMultipleMergedRegionsProcedure is still going on, in this case, the second
+    // GCMultipleMergedRegionsProcedure could delete the merged region by accident!
+    if (qualifiers.isEmpty()) {
+      LOG.info("No merged qualifiers for region " + mergeRegion.getRegionNameAsString() +
+        " in meta table, they are cleaned up already, Skip.");
+      return;
+    }
+    try (Table table = master.getConnection().getTable(TableName.META_TABLE_NAME)) {
+      table.delete(delete);
+    }
+    LOG.info("Deleted merge references in " + mergeRegion.getRegionNameAsString() +
+      ", deleted qualifiers " +
+      qualifiers.stream().map(Bytes::toStringBinary).collect(Collectors.joining(", ")));
+  }
+
+  @VisibleForTesting
+  static Put addMergeRegions(Put put, Collection<RegionInfo> mergeRegions) throws IOException {
+    int limit = 10000; // Arbitrary limit. No room in our formatted 'task0000' below for more.
+    int max = mergeRegions.size();
+    if (max > limit) {
+      // Should never happen!!!!! But just in case.
+      throw new RuntimeException(
+        "Can't merge " + max + " regions in one go; " + limit + " is upper-limit.");
+    }
+    int counter = 0;
+    for (RegionInfo ri : mergeRegions) {
+      String qualifier = String.format(HConstants.MERGE_QUALIFIER_PREFIX_STR + "%04d", counter++);
+      put.add(CellBuilderFactory.create(CellBuilderType.SHALLOW_COPY).setRow(put.getRow())
+        .setFamily(HConstants.CATALOG_FAMILY).setQualifier(Bytes.toBytes(qualifier))
+        .setTimestamp(put.getTimestamp()).setType(Type.Put).setValue(RegionInfo.toByteArray(ri))
+        .build());
+    }
+    return put;
   }
 
   // ============================================================================================
@@ -333,5 +549,20 @@ public class RegionStateStore {
         ? HConstants.STATE_QUALIFIER
         : Bytes.toBytes(HConstants.STATE_QUALIFIER_STR + META_REPLICA_ID_DELIMITER
           + String.format(RegionInfo.REPLICA_ID_FORMAT, replicaId));
+  }
+
+  private static void debugLogMutations(List<? extends Mutation> mutations) throws IOException {
+    if (!METALOG.isDebugEnabled()) {
+      return;
+    }
+    // Logging each mutation in separate line makes it easier to see diff between them visually
+    // because of common starting indentation.
+    for (Mutation mutation : mutations) {
+      debugLogMutation(mutation);
+    }
+  }
+
+  private static void debugLogMutation(Mutation p) throws IOException {
+    METALOG.debug("{} {}", p.getClass().getSimpleName(), p.toJSON());
   }
 }
