+++ hbase-server/src/test/java/org/apache/hadoop/hbase/master/assignment/TestRegionStateStore.java
@@ -17,34 +17,52 @@
  */
 package org.apache.hadoop.hbase.master.assignment;
 
+import static org.apache.hadoop.hbase.TestMetaTableAccessor.assertEmptyMetaLocation;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertNotEquals;
+import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
+
+import java.io.IOException;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.concurrent.ThreadLocalRandom;
+import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
+import org.apache.hadoop.hbase.CatalogFamilyFormat;
+import org.apache.hadoop.hbase.Cell;
 import org.apache.hadoop.hbase.HBaseClassTestRule;
 import org.apache.hadoop.hbase.HBaseTestingUtility;
 import org.apache.hadoop.hbase.HConstants;
+import org.apache.hadoop.hbase.MetaTableAccessor;
 import org.apache.hadoop.hbase.ServerName;
 import org.apache.hadoop.hbase.TableName;
+import org.apache.hadoop.hbase.TableNameTestRule;
+import org.apache.hadoop.hbase.client.Admin;
+import org.apache.hadoop.hbase.client.Get;
 import org.apache.hadoop.hbase.client.Put;
 import org.apache.hadoop.hbase.client.RegionInfo;
+import org.apache.hadoop.hbase.client.RegionInfoBuilder;
 import org.apache.hadoop.hbase.client.Result;
 import org.apache.hadoop.hbase.client.Table;
+import org.apache.hadoop.hbase.client.TableDescriptorBuilder;
 import org.apache.hadoop.hbase.master.RegionState;
 import org.apache.hadoop.hbase.regionserver.HRegion;
 import org.apache.hadoop.hbase.testclassification.MasterTests;
 import org.apache.hadoop.hbase.testclassification.MediumTests;
 import org.apache.hadoop.hbase.util.Bytes;
 import org.apache.hadoop.hbase.util.EnvironmentEdgeManager;
+import org.apache.hadoop.hbase.util.ManualEnvironmentEdge;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.ClassRule;
+import org.junit.Rule;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+
+import org.apache.hbase.thirdparty.com.google.common.collect.Lists;
 
 
 @Category({ MasterTests.class, MediumTests.class })
@@ -54,9 +72,10 @@ public class TestRegionStateStore {
   public static final HBaseClassTestRule CLASS_RULE =
     HBaseClassTestRule.forClass(TestRegionStateStore.class);
 
-  private static final Logger LOG = LoggerFactory.getLogger(TestRegionStateStore.class);
-
   private static HBaseTestingUtility UTIL = new HBaseTestingUtility();
+
+  @Rule
+  public final TableNameTestRule name = new TableNameTestRule();
 
   @BeforeClass
   public static void beforeClass() throws Exception {
@@ -135,5 +154,244 @@ public class TestRegionStateStore {
       }
     });
     assertFalse("Visitor has been called, but it shouldn't.", visitorCalled.get());
+  }
+
+  @Test
+  public void testMetaLocationForRegionReplicasIsAddedAtRegionSplit() throws IOException {
+    long regionId = System.currentTimeMillis();
+    ServerName serverName0 =
+      ServerName.valueOf("foo", 60010, ThreadLocalRandom.current().nextLong());
+    TableName tableName = name.getTableName();
+    RegionInfo parent = RegionInfoBuilder.newBuilder(tableName)
+      .setStartKey(HConstants.EMPTY_START_ROW).setEndKey(HConstants.EMPTY_END_ROW).setSplit(false)
+      .setRegionId(regionId).setReplicaId(0).build();
+
+    RegionInfo splitA = RegionInfoBuilder.newBuilder(tableName)
+      .setStartKey(HConstants.EMPTY_START_ROW).setEndKey(Bytes.toBytes("a")).setSplit(false)
+      .setRegionId(regionId + 1).setReplicaId(0).build();
+    RegionInfo splitB = RegionInfoBuilder.newBuilder(tableName).setStartKey(Bytes.toBytes("a"))
+      .setEndKey(HConstants.EMPTY_END_ROW).setSplit(false).setRegionId(regionId + 1).setReplicaId(0)
+      .build();
+    List<RegionInfo> regionInfos = Lists.newArrayList(parent);
+    MetaTableAccessor.addRegionsToMeta(UTIL.getConnection(), regionInfos, 3);
+    final RegionStateStore regionStateStore =
+      UTIL.getHBaseCluster().getMaster().getAssignmentManager().getRegionStateStore();
+    regionStateStore.splitRegion(parent, splitA, splitB, serverName0,
+      TableDescriptorBuilder.newBuilder(tableName).setRegionReplication(3).build());
+    try (Table meta = MetaTableAccessor.getMetaHTable(UTIL.getConnection())) {
+      assertEmptyMetaLocation(meta, splitA.getRegionName(), 1);
+      assertEmptyMetaLocation(meta, splitA.getRegionName(), 2);
+      assertEmptyMetaLocation(meta, splitB.getRegionName(), 1);
+      assertEmptyMetaLocation(meta, splitB.getRegionName(), 2);
+    }
+  }
+
+  @Test
+  public void testEmptyMetaDaughterLocationDuringSplit() throws IOException {
+    TableName tableName = name.getTableName();
+    long regionId = System.currentTimeMillis();
+    ServerName serverName0 =
+      ServerName.valueOf("foo", 60010, ThreadLocalRandom.current().nextLong());
+    RegionInfo parent = RegionInfoBuilder.newBuilder(tableName)
+      .setStartKey(HConstants.EMPTY_START_ROW).setEndKey(HConstants.EMPTY_END_ROW).setSplit(false)
+      .setRegionId(regionId).setReplicaId(0).build();
+    RegionInfo splitA = RegionInfoBuilder.newBuilder(tableName)
+      .setStartKey(HConstants.EMPTY_START_ROW).setEndKey(Bytes.toBytes("a")).setSplit(false)
+      .setRegionId(regionId + 1).setReplicaId(0).build();
+    RegionInfo splitB = RegionInfoBuilder.newBuilder(tableName).setStartKey(Bytes.toBytes("a"))
+      .setEndKey(HConstants.EMPTY_END_ROW).setSplit(false).setRegionId(regionId + 1).setReplicaId(0)
+      .build();
+    List<RegionInfo> regionInfos = Lists.newArrayList(parent);
+    MetaTableAccessor.addRegionsToMeta(UTIL.getConnection(), regionInfos, 3);
+    final RegionStateStore regionStateStore =
+      UTIL.getHBaseCluster().getMaster().getAssignmentManager().getRegionStateStore();
+    regionStateStore.splitRegion(parent, splitA, splitB, serverName0,
+      TableDescriptorBuilder.newBuilder(tableName).setRegionReplication(3).build());
+    try (Table meta = MetaTableAccessor.getMetaHTable(UTIL.getConnection())) {
+      Get get1 = new Get(splitA.getRegionName());
+      Result resultA = meta.get(get1);
+      Cell serverCellA = resultA.getColumnLatestCell(HConstants.CATALOG_FAMILY,
+        CatalogFamilyFormat.getServerColumn(splitA.getReplicaId()));
+      Cell startCodeCellA = resultA.getColumnLatestCell(HConstants.CATALOG_FAMILY,
+        CatalogFamilyFormat.getStartCodeColumn(splitA.getReplicaId()));
+      assertNull(serverCellA);
+      assertNull(startCodeCellA);
+
+      Get get2 = new Get(splitA.getRegionName());
+      Result resultB = meta.get(get2);
+      Cell serverCellB = resultB.getColumnLatestCell(HConstants.CATALOG_FAMILY,
+        CatalogFamilyFormat.getServerColumn(splitB.getReplicaId()));
+      Cell startCodeCellB = resultB.getColumnLatestCell(HConstants.CATALOG_FAMILY,
+        CatalogFamilyFormat.getStartCodeColumn(splitB.getReplicaId()));
+      assertNull(serverCellB);
+      assertNull(startCodeCellB);
+    }
+  }
+
+  @Test
+  public void testMetaLocationForRegionReplicasIsAddedAtRegionMerge() throws IOException {
+    long regionId = System.currentTimeMillis();
+    ServerName serverName0 =
+      ServerName.valueOf("foo", 60010, ThreadLocalRandom.current().nextLong());
+
+    TableName tableName = name.getTableName();
+    RegionInfo parentA = RegionInfoBuilder.newBuilder(tableName).setStartKey(Bytes.toBytes("a"))
+      .setEndKey(HConstants.EMPTY_END_ROW).setSplit(false).setRegionId(regionId).setReplicaId(0)
+      .build();
+
+    RegionInfo parentB = RegionInfoBuilder.newBuilder(tableName)
+      .setStartKey(HConstants.EMPTY_START_ROW).setEndKey(Bytes.toBytes("a")).setSplit(false)
+      .setRegionId(regionId).setReplicaId(0).build();
+    RegionInfo merged = RegionInfoBuilder.newBuilder(tableName)
+      .setStartKey(HConstants.EMPTY_START_ROW).setEndKey(HConstants.EMPTY_END_ROW).setSplit(false)
+      .setRegionId(regionId + 1).setReplicaId(0).build();
+
+    final RegionStateStore regionStateStore =
+      UTIL.getHBaseCluster().getMaster().getAssignmentManager().getRegionStateStore();
+
+    try (Table meta = MetaTableAccessor.getMetaHTable(UTIL.getConnection())) {
+      List<RegionInfo> regionInfos = Lists.newArrayList(parentA, parentB);
+      MetaTableAccessor.addRegionsToMeta(UTIL.getConnection(), regionInfos, 3);
+      regionStateStore.mergeRegions(merged, new RegionInfo[] { parentA, parentB }, serverName0,
+        TableDescriptorBuilder.newBuilder(tableName).setRegionReplication(3).build());
+      assertEmptyMetaLocation(meta, merged.getRegionName(), 1);
+      assertEmptyMetaLocation(meta, merged.getRegionName(), 2);
+    }
+  }
+
+  @Test
+  public void testMastersSystemTimeIsUsedInMergeRegions() throws IOException {
+    long regionId = System.currentTimeMillis();
+    TableName tableName = name.getTableName();
+
+    RegionInfo regionInfoA = RegionInfoBuilder.newBuilder(tableName)
+      .setStartKey(HConstants.EMPTY_START_ROW).setEndKey(new byte[] { 'a' }).setSplit(false)
+      .setRegionId(regionId).setReplicaId(0).build();
+
+    RegionInfo regionInfoB = RegionInfoBuilder.newBuilder(tableName).setStartKey(new byte[] { 'a' })
+      .setEndKey(HConstants.EMPTY_END_ROW).setSplit(false).setRegionId(regionId).setReplicaId(0)
+      .build();
+    RegionInfo mergedRegionInfo = RegionInfoBuilder.newBuilder(tableName)
+      .setStartKey(HConstants.EMPTY_START_ROW).setEndKey(HConstants.EMPTY_END_ROW).setSplit(false)
+      .setRegionId(regionId).setReplicaId(0).build();
+
+    ServerName sn = ServerName.valueOf("bar", 0, 0);
+    try (Table meta = MetaTableAccessor.getMetaHTable(UTIL.getConnection())) {
+      List<RegionInfo> regionInfos = Lists.newArrayList(regionInfoA, regionInfoB);
+      MetaTableAccessor.addRegionsToMeta(UTIL.getConnection(), regionInfos, 1);
+
+      // write the serverName column with a big current time, but set the masters time as even
+      // bigger. When region merge deletes the rows for regionA and regionB, the serverName columns
+      // should not be seen by the following get
+      long serverNameTime = EnvironmentEdgeManager.currentTime() + 100000000;
+      long masterSystemTime = EnvironmentEdgeManager.currentTime() + 123456789;
+
+      // write the serverName columns
+      MetaTableAccessor.updateRegionLocation(UTIL.getConnection(), regionInfoA, sn, 1,
+        serverNameTime);
+
+      // assert that we have the serverName column with expected ts
+      Get get = new Get(mergedRegionInfo.getRegionName());
+      Result result = meta.get(get);
+      Cell serverCell = result.getColumnLatestCell(HConstants.CATALOG_FAMILY,
+        CatalogFamilyFormat.getServerColumn(0));
+      assertNotNull(serverCell);
+      assertEquals(serverNameTime, serverCell.getTimestamp());
+
+      final RegionStateStore regionStateStore =
+        UTIL.getHBaseCluster().getMaster().getAssignmentManager().getRegionStateStore();
+
+      ManualEnvironmentEdge edge = new ManualEnvironmentEdge();
+      edge.setValue(masterSystemTime);
+      EnvironmentEdgeManager.injectEdge(edge);
+      try {
+        // now merge the regions, effectively deleting the rows for region a and b.
+        regionStateStore.mergeRegions(mergedRegionInfo,
+          new RegionInfo[] { regionInfoA, regionInfoB }, sn,
+          TableDescriptorBuilder.newBuilder(tableName).build());
+      } finally {
+        EnvironmentEdgeManager.reset();
+      }
+
+      result = meta.get(get);
+      serverCell = result.getColumnLatestCell(HConstants.CATALOG_FAMILY,
+        CatalogFamilyFormat.getServerColumn(0));
+      Cell startCodeCell = result.getColumnLatestCell(HConstants.CATALOG_FAMILY,
+        CatalogFamilyFormat.getStartCodeColumn(0));
+      Cell seqNumCell = result.getColumnLatestCell(HConstants.CATALOG_FAMILY,
+        CatalogFamilyFormat.getSeqNumColumn(0));
+      assertNull(serverCell);
+      assertNull(startCodeCell);
+      assertNull(seqNumCell);
+    }
+  }
+
+  /**
+   * Test for HBASE-23044.
+   */
+  @Test
+  public void testGetMergeRegions() throws Exception {
+    TableName tn = name.getTableName();
+    UTIL.createMultiRegionTable(tn, Bytes.toBytes("CF"), 4);
+    UTIL.waitTableAvailable(tn);
+    Admin admin = UTIL.getAdmin();
+    List<RegionInfo> regions = admin.getRegions(tn);
+    assertEquals(4, regions.size());
+    admin
+      .mergeRegionsAsync(
+        new byte[][] { regions.get(0).getRegionName(), regions.get(1).getRegionName() }, false)
+      .get(60, TimeUnit.SECONDS);
+    admin
+      .mergeRegionsAsync(
+        new byte[][] { regions.get(2).getRegionName(), regions.get(3).getRegionName() }, false)
+      .get(60, TimeUnit.SECONDS);
+
+    List<RegionInfo> mergedRegions = admin.getRegions(tn);
+    assertEquals(2, mergedRegions.size());
+    RegionInfo mergedRegion0 = mergedRegions.get(0);
+    RegionInfo mergedRegion1 = mergedRegions.get(1);
+
+    final RegionStateStore regionStateStore =
+      UTIL.getHBaseCluster().getMaster().getAssignmentManager().getRegionStateStore();
+
+    List<RegionInfo> mergeParents = regionStateStore.getMergeRegions(mergedRegion0);
+    assertTrue(mergeParents.contains(regions.get(0)));
+    assertTrue(mergeParents.contains(regions.get(1)));
+    mergeParents = regionStateStore.getMergeRegions(mergedRegion1);
+    assertTrue(mergeParents.contains(regions.get(2)));
+    assertTrue(mergeParents.contains(regions.get(3)));
+
+    // Delete merge qualifiers for mergedRegion0, then cannot getMergeRegions again
+    regionStateStore.deleteMergeQualifiers(mergedRegion0);
+    mergeParents = regionStateStore.getMergeRegions(mergedRegion0);
+    assertNull(mergeParents);
+
+    mergeParents = regionStateStore.getMergeRegions(mergedRegion1);
+    assertTrue(mergeParents.contains(regions.get(2)));
+    assertTrue(mergeParents.contains(regions.get(3)));
+  }
+
+  @Test
+  public void testAddMergeRegions() throws IOException {
+    TableName tn = name.getTableName();
+    Put put = new Put(Bytes.toBytes(name.getTableName().getNameAsString()));
+    List<RegionInfo> ris = new ArrayList<>();
+    int limit = 10;
+    byte[] previous = HConstants.EMPTY_START_ROW;
+    for (int i = 0; i < limit; i++) {
+      RegionInfo ri =
+        RegionInfoBuilder.newBuilder(tn).setStartKey(previous).setEndKey(Bytes.toBytes(i)).build();
+      ris.add(ri);
+    }
+    put = RegionStateStore.addMergeRegions(put, ris);
+    List<Cell> cells = put.getFamilyCellMap().get(HConstants.CATALOG_FAMILY);
+    String previousQualifier = null;
+    assertEquals(limit, cells.size());
+    for (Cell cell : cells) {
+      String qualifier = Bytes.toString(cell.getQualifierArray());
+      assertTrue(qualifier.startsWith(HConstants.MERGE_QUALIFIER_PREFIX_STR));
+      assertNotEquals(qualifier, previousQualifier);
+      previousQualifier = qualifier;
+    }
   }
 }
